requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.16.0
idna==2.7
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.19.1
six==1.11.0
urllib3==1.23
